refactor: split utils files to their own files
