+++ src/strategy/utils/decodeIPFSStrategyPayload.ts
@@ -1,0 +1,6 @@
+import { IPFSStrategyPayload, IPFSStrategyPayloadSchema } from '../../types';
+
+export const decodeIPFSStrategyPayload = (
+  ipfsStrategyPayload: string
+): IPFSStrategyPayload =>
+  IPFSStrategyPayloadSchema.parse(JSON.parse(ipfsStrategyPayload));
